separate substrate type and supporting routine
