+++ fearless/Common/Substrate/Types/BalanceLock.swift
@@ -3,8 +3,6 @@ import BigInt
 import FearlessUtils
 import SoraFoundation
 
-typealias BalanceLocks = [BalanceLock]
-
 struct BalanceLock: Codable, Equatable {
     enum CodingKeys: String, CodingKey {
         case identifier = "id"
@@ -24,57 +22,8 @@ struct BalanceLock: Codable, Equatable {
     }
 }
 
-extension BalanceLocks {
-    func mainLocks() -> BalanceLocks {
-        LockType.locksOrder.compactMap { lockType in
-            self.first(where: { lock in
-                lock.displayId == lockType.rawValue
-            })
-        }
-    }
-
-    func auxLocks() -> BalanceLocks {
-        compactMap { lock in
-            guard LockType(rawValue: lock.displayId ?? "") != nil else {
-                return lock
-            }
-
-            return nil
-        }.sorted { lhs, rhs in
-            lhs.amount > rhs.amount
-        }
-    }
-}
-
 enum LockReason: UInt8, Codable {
     case fee
     case misc
     case all
-}
-
-enum LockType: String {
-    case staking
-    case vesting
-    case democracy = "democrac"
-
-    static var locksOrder: [Self] = [.vesting, .staking, .democracy]
-
-    var displayType: LocalizableResource<String> {
-        LocalizableResource<String> { locale in
-            switch self {
-            case .vesting:
-                return R.string.localizable.walletAccountLocksVesting(
-                    preferredLanguages: locale.rLanguages
-                )
-            case .staking:
-                return R.string.localizable.stakingTitle(
-                    preferredLanguages: locale.rLanguages
-                )
-            case .democracy:
-                return R.string.localizable.walletAccountLocksDemocracy(
-                    preferredLanguages: locale.rLanguages
-                )
-            }
-        }
-    }
 }
